cliente-pedido-service/airflow/requirements.txt
# airflow/requirements.txt
kafka-python==2.0.2
requests==2.31.0
